API_Flask/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Limiter==1.4
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
limits==1.5.1
MarkupSafe==1.1.1
PyJWT==1.4.2
six==1.15.0
Werkzeug==1.0.1
